bip0039::language
Trait Language

pub trait Language: Sized {
    const WORD_LIST: &'static [&'static str];

    // Provided methods
    fn word_of(index: usize) -> &'static str { ... }
    fn index_of(word: &str) -> Option<usize> { ... }
    fn is_sorted() -> bool { ... }
    fn words_by_prefix(prefix: &str) -> &[&'static str] { ... }
}

Language to be used for the mnemonic phrase.

The English language is always available, other languages are enabled using the compilation features.

Required Associated Constants

const WORD_LIST: &'static [&'static str]

The word list for this language.

Provided Methods

fn word_of(index: usize) -> &'static str

Returns the word of index in the word list.

fn index_of(word: &str) -> Option<usize>

Returns the index of the word in the word list.

fn is_sorted() -> bool

Checks if the word list of this language are sorted by the byte values.

The words in the word list are ordered lexicographically, which means that we cannot use binary_search to find words more efficiently if the ordering of the words is not be sorted by the byte values, because the Rust ordering is based on the byte values.

fn words_by_prefix(prefix: &str) -> &[&'static str]

Returns words from the word list that start with the given prefix.

The words in the word list are ordered lexicographically, which means that we cannot use binary_search to find words more efficiently if the ordering of the words is not be sorted by the byte values, because the Rust ordering is based on the byte values. However, it does mean that words that share a prefix will follow each other.

Object Safety

This trait is not object safe.